+++ src/memo/memo_intf.ml
@@ -7,6 +7,13 @@ module type Data = sig
   val to_sexp : t -> Sexp.t
 end
 
+module type Input = sig
+  include Data
+  val decode : t Dune_lang.Decoder.t
+end
+
+module type Output = Data
+
 module type S = sig
   type input
 
@@ -27,7 +34,8 @@ module type S = sig
   val create
     :  string
     -> ?allow_cutoff:bool
-    -> (module Data with type t = 'a)
+    -> doc:string
+    -> (module Output with type t = 'a)
     -> (input -> 'a Fiber.t)
     -> 'a t